requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.11.0
python-docx==1.1.0
spacy==3.7.2
werkzeug==3.0.1
